[measureobjectskeleton]: (3) move code from frontend/module to library/module, created new opts for module-specific constants, combined code for skeleton graph and branchpoint image into the same function call
